requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
python-dateutil==2.9.0
